--- a/datalad/core/local/create.py
+++ b/datalad/core/local/create.py
@@ -389,7 +389,7 @@ class Create(Interface):
         # and unnecessary for the Python API (there could simply be a
         # subsequence ds.config.add() call)
         for k, v in tbds_config.overrides.items():
-            tbds_config.add(k, v, where='local', reload=False)
+            tbds_config.add(k, v, scope='local', reload=False)
 
         # all config manipulation is done -> fll reload
         tbds_config.reload()