… benchmarks/plugins/addurls.py
… datalad/core/distributed/clone.py
… datalad/core/distributed/tests/test_clone.py
… datalad/core/distributed/tests/test_push.py
… datalad/core/local/tests/test_resulthooks.py
… datalad/core/local/tests/test_run.py
… datalad/core/local/tests/test_save.py
… datalad/distributed/create_sibling_gin.py
… datalad/distributed/create_sibling_gitlab.py
… datalad/distributed/create_sibling_ria.py
… datalad/distributed/tests/test_create_sibling_ria.py
… datalad/distributed/tests/test_ria_basics.py
… datalad/distribution/get.py
… datalad/distribution/siblings.py
… datalad/distribution/tests/test_get.py
… datalad/distribution/tests/test_siblings.py
… datalad/distribution/tests/test_update.py
… datalad/interface/tests/test_rerun.py
… datalad/interface/tests/test_run_procedure.py
… datalad/metadata/extractors/tests/test_audio.py
… datalad/metadata/extractors/tests/test_exif.py
… datalad/metadata/extractors/tests/test_image.py
… datalad/metadata/extractors/tests/test_xmp.py
… datalad/metadata/tests/test_base.py
… datalad/metadata/tests/test_search.py
… datalad/resources/procedures/cfg_metadatatypes.py
… datalad/support/annexrepo.py
… datalad/support/gitrepo.py
… datalad/support/tests/test_annexrepo.py
… datalad/support/tests/test_gitrepo.py