A follow up to the call https://github.com/datalad/datalad/issues/5489#issuecomment-915404977 -- as I see it could potentially be done

- [x] datalad.tests.test_config pass
- [x] decide on the destiny of `source` which underlies where/scope
- [x] comb through docstrings/comments on what to revert back to use word "where"
- [x] investigate if there is any notable effect on performance due to decoration/checks
- [x] extend tests to check (and likely fix) deprecated usage still working out

note: initial replacements are done using `datalad run` so in principle this PR might be `datalad rerun`able if we start breeding conflicts... would be fun to check -- is opening up a DX usage scenario for `run/rerun`.