@mih 2021-09-14
re peformance; it seems to be kosher:

Using `timeit cfg.set('mike.this', 'value', where='global')`

```py
# with decorator, old call
>>> timeit cfg.set('mike.this', 'value', where='global')
10.4 ms ± 202 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

# with decorator, new call
timeit cfg.set('mike.this', 'value', scope='global')
10.1 ms ± 45.3 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

# without decorator, new call
timeit cfg.set('mike.this', 'value', scope='global')
9.96 ms ± 10.9 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

IOW: <5% performance slowdown for old unchanged code, <2% in comparison to a hard transition (breakage) to a new API.

Given that only *write* access is affected, this is tolerable IMHO.