@mih 2022-01-28
> @mih -- we should either strive to adhere to pep8 from now on (e.g. in diffs at least; may be with a quick pep8-ification maraphone after 0.16.0) or disable pep8 checks on codeclimate. What would you prefer?

I don't care. I consider this really minor in comparison to all other issues. But I know that without a check being done, nothing will ever change.