@yarikoptic 2022-02-08
@mih:
> Hence I will not object this PR, if this is the only way forward.

There are always choices, e.g. here "merge" or "close". Since no stronger objection against "merge", I will take it.

@bpoldrack :
> Ultimately I am OK with a merge.

Coolio, thank you.

Overall, I think such harmonization is benefitial. Due to deprecation compatibility layer it should be not effecting users much, even if some rare user uses `scope` explicitly.  So I feel that it is a good way forward regardless either we ever see fusion of `config` and `configuration` or "privatization" of `config`.  Let's proceed so it doesn't hang over any longer.